Refactor parse of category query params
